requirements.txt
fastapi
uvicorn==0.11.8
uvloop==0.14.0
requests
beautifulsoup4==4.9.3
lxml
readability-lxml
PyPDF2==1.26.0
rank-bm25
jieba
